* ReSpawn 시 플레이어 깜빡이면서 스폰되도록 변경 * Text UI 구현. String 텍스트와 Text Size를 입력받아 문자별로 이미지를 매칭시켜 String 이미지 출력 * Life UI 구현. Player의 남은 Life 만큼 플레이어 이미지를 출력

+++ ShootingStrike/HitEffect.cpp
@@ -19,7 +19,7 @@ void HitEffect::Initialize()
 	Super::Initialize();
 
 	if ( pImage )
-		segmentImageCount = pImage->GetScale().x / pImage->GetSegmentationScale().x;
+		segmentImageCount = static_cast<int>(pImage->GetScale().x / pImage->GetSegmentationScale().x);
 	else
 		segmentImageCount = 1;
